coalesce {microbenchmark}	R Documentation

Return first non null argument.

Description

This function is useful when processing complex arguments with multiple possible defaults based on other arguments that may or may not have been provided.

Usage

coalesce(...)

Arguments

...
List of values.

Value

First non null element in ....

Author(s)

Olaf Mersmann

[Package microbenchmark version 1.5.0 Index]