convert_to_unit {microbenchmark}	R Documentation

Convert timings to different units.

Description

The following units of time are supported

“ns”

Nanoseconds.

“us”

Microseconds.

“ms”

Milliseconds.

“s”

Seconds.

“t”

Appropriately prefixed time unit.

“hz”

Hertz / evaluations per second.

“eps”

Evaluations per second / Hertz.

“khz”

Kilohertz / 1000s of evaluations per second.

“mhz”

Megahertz / 1000000s of evaluations per second.

“f”

Appropriately prefixed frequency unit.

Usage

convert_to_unit(object, unit)

Arguments

object
A microbenchmark object.

unit
A unit of time. See details.

Value

A matrix containing the converted time values with an attribute unit which is a printable name of the unit of time.

Author(s)

Olaf Mersmann

[Package microbenchmark version 1.5.0 Index]