get_nanotime {microbenchmark}	R Documentation

Return the current value of the platform timer.

Description

The current value of the most accurate timer of the platform is returned. This can be used as a time stamp for logging or similar purposes. Please note that there is no common reference, that is, the timer value cannot be converted to a date and time value.

Usage

get_nanotime()

Author(s)

Olaf Mersmann

[Package microbenchmark version 1.5.0 Index]