microtiming_precision {microbenchmark}	R Documentation

Estimate precision of timing routines.

Description

This function is currently experimental. Its main use is to judge the quality of the underlying timer implementation of the operating system. The function measures the overhead of timing a C function call rounds times and returns all non-zero timings observed. This can be used to judge the granularity and resolution of the timing subsystem.

Usage

microtiming_precision(rounds = 100L, warmup = 2^18)

Arguments

rounds
Number of measurements used to estimate the precision.

warmup
Number of iterations used to warm up the CPU.

Value

A vector of observed non-zero timings.

Author(s)

Olaf Mersmann

[Package microbenchmark version 1.5.0 Index]