print.microbenchmark {microbenchmark}	R Documentation

Print microbenchmark timings.

Description

Print microbenchmark timings.

Usage

## S3 method for class 'microbenchmark'
print(x, unit, order, signif, ...)

Arguments

x
An object of class microbenchmark.

unit
What unit to print the timings in. Default value taken from to option microbenchmark.unit (see example).

order
If present, order results according to this column of the output.

signif
If present, limit the number of significant digits shown.

...
Passed to print.data.frame

Note

The available units are nanoseconds ("ns"), microseconds ("us"), milliseconds ("ms"), seconds ("s") and evaluations per seconds ("eps") and relative runtime compared to the best median time ("relative").

If the multcomp package is available a statistical ranking is calculated and displayed in compact letter display from in the cld column.

Author(s)

Olaf Mersmann

See Also

boxplot.microbenchmark and autoplot.microbenchmark for a plot methods.

Examples

a1 <- a2 <- a3 <- a4 <- numeric(0)

res <- microbenchmark(a1 <- c(a1, 1),
                      a2 <- append(a2, 1),
                      a3[length(a3) + 1] <- 1,
                      a4[[length(a4) + 1]] <- 1,
                      times=100L)
print(res)
## Change default unit to relative runtime
options(microbenchmark.unit="relative")
print(res)
## Change default unit to evaluations per second
options(microbenchmark.unit="eps")
print(res)

[Package microbenchmark version 1.5.0 Index]