summary.microbenchmark {microbenchmark}	R Documentation

Summarize microbenchmark timings.

Description

Summarize microbenchmark timings.

Usage

## S3 method for class 'microbenchmark'
summary(object, unit, ..., include_cld = TRUE)

Arguments

object
An object of class microbenchmark.

unit
What unit to print the timings in. If none is given, either the unit attribute of object or the option microbenchmark.unit is used and if neither is set “t” is used.

...
Ignored

include_cld
Calculate cld using multcomp::glht() and add it to the output. Set to FALSE if the calculation takes too long.

Value

A data.frame containing the aggregated results.

Note

The available units are nanoseconds ("ns"), microseconds ("us"), milliseconds ("ms"), seconds ("s") and evaluations per seconds ("eps") and relative runtime compared to the best median time ("relative").

See Also

print.microbenchmark

[Package microbenchmark version 1.5.0 Index]